move slash input
